Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/xhr/no-utf16-json-expected.txt


FAIL JSON asa UTF-16 with BOM should decode as UTF-8 (and, therefore, fail) assert_equals: expected null but got object "[object Object]"
